Multimodule - using calculator in main method

+++ src/main/java/khan/riaz/mohammed/CalculatorInterface.java
@@ -12,7 +12,7 @@ public interface CalculatorInterface {
 
     int modulus(int a, int b);
 
-    int percent(int a);
+    double percent(double a);
 
 
 }
